refactor(worker): extract dispatcher cache into src/worker/dispatcher_cache
