+++ Makefile
@@ -38,6 +38,7 @@ OBJS += src/utils/logging.o
 OBJS += src/utils/cloudevents.o
 OBJS += src/utils/rate_limit.o
 OBJS += src/utils/retry_policy.o
+OBJS += src/worker/dispatcher_cache.o
 
 # Conditional protocol objects
 ifdef ENABLE_KAFKA
